+++ src/main/java/kitchenpos/application/ProductService.java
@@ -4,7 +4,7 @@ import java.util.stream.Collectors;
 import kitchenpos.application.dto.ProductCreateRequest;
 import kitchenpos.application.dto.ProductResponse;
 import kitchenpos.domain.product.Product;
-import kitchenpos.repository.ProductRepository;
+import kitchenpos.domain.product.ProductRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
