+++ src/main/java/kitchenpos/domain/table/TableGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.repository;
+package kitchenpos.domain.table;
 
 import kitchenpos.domain.table.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
